# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZK-VOTE Relayer API",
    "version": "1.0.0",
    "description": "Backend relayer for anonymous voting on Stellar/Soroban. Generated from route definitions and Zod validation schemas — see backend/API.md for prose docs and GET /api-docs for interactive documentation."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "RelayerAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Relayer-Auth",
        "description": "Shared relayer auth token (also accepted as `Authorization: Bearer <token>`). There is currently one shared token per deployment, not per-user credentials."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Unauthorized"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "DaosListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {}
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "cursor": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "hasMore": {
                "type": "boolean"
              },
              "total": {
                "type": "number"
              }
            },
            "required": [
              "hasMore",
              "total"
            ]
          },
          "lastSync": {
            "type": [
              "string",
              "null"
            ]
          },
          "cached": {
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "pagination",
          "lastSync",
          "cached"
        ]
      }
    },
...
  },
  "paths": {
...
    "/daos": {
      "get": {
        "tags": [
          "DAOs"
        ],
        "summary": "List cached DAOs with pagination, optionally including user membership role",
        "description": "Rate limit: queryLimiter.",
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Stellar address",
              "example": "GABCDEF..."
            },
            "required": false,
            "name": "user",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "description": "Page size (default 100)",
              "example": 100
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Rows to skip (default 0)",
              "example": 0
            },
            "required": false,
            "name": "offset",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Opaque cursor from a previous pagination.cursor; alias for offset",
              "example": "100"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaosListResponse"
                },
                "example": {
                  "data": [],
                  "pagination": {
                    "hasMore": false,
                    "total": 0
                  },
                  "lastSync": null,
                  "cached": true
                }
              }
            }
          },
          "400": {
            "description": "Error (HTTP 400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error (HTTP 500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/dao/{daoId}": {
      "get": {
        "tags": [
          "DAOs"
        ],
        "summary": "Get a single cached DAO",
        "description": "Rate limit: queryLimiter.",
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "DAO identifier",
              "example": "0"
            },
            "required": true,
            "name": "daoId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "dao": {
                    "id": 0,
                    "name": "Example DAO",
                    "creator": "GABCDEF..."
                  },
                  "cached": true
                }
              }
            }
          },
          "404": {
            "description": "Error (HTTP 404)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/daos/sync": {
      "post": {
        "tags": [
          "DAOs"
        ],
        "summary": "Trigger a manual DAO sync from the on-chain registry",
        "description": "No rate limit.",
        "security": [
          {
            "RelayerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "success": true,
                  "synced": 5
                }
              }
            }
          },
          "401": {
            "description": "Error (HTTP 401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error (HTTP 500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}